requirements.txt
dj-database-url==0.3.0
Django==1.7
django-allauth==0.19.0
django-cors-headers==1.0.0
django-guardian==1.2.5
django-rest-auth==0.3.4
djangorestframework==3.0.4
httpie==0.9.0
oauthlib==0.7.2
Pillow==2.7.0
Pygments==2.0.2
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
waitress==0.8.9
whitenoise==1.0.6
psycopg2==2.5.4